Expiration Complete Listener in Order service (class and tests) added

+++ tickets/src/models/ticket.ts
@@ -21,7 +21,7 @@ export class Ticket extends BaseEntity {
   @Column({ nullable: false })
   userId: number;
 
-  @Column({ type: "numeric", nullable: true })
+  @Column({ type: "int", nullable: true })
   orderId: number | null;
 
   @VersionColumn()
